refactors result approvals in separate package
